switch to bigint

+++ test/support/migrations.exs
@@ -7,7 +7,7 @@ defmodule GuardianDb.Test.Repo.Migrations do
       add :aud, :string
       add :iss, :string
       add :sub, :string
-      add :exp, :integer
+      add :exp, :bigint
       add :jwt, :text
       add :claims, :text
